refactor: modularize llmsTxtPlugin and simplify docusaurus.config.ts
